backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# WebSocket support
websockets==12.0

# Database
sqlalchemy[asyncio]==2.0.23
aiosqlite==0.19.0
alembic==1.12.1

# HTTP client
aiohttp==3.9.1

# Image processing
Pillow==10.1.0

# Scientific computing
numpy==1.25.2
scipy==1.11.4

# Signal processing for EEG
matplotlib==3.8.2
scikit-learn==1.3.2

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging and monitoring
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Optional: Real EMOTIV SDK (uncomment if using real device)
# cortex-python==1.0.0

# Optional: Additional ML libraries for advanced confusion detection
# tensorflow==2.15.0
# torch==2.1.0

# CORS support
python-multipart==0.0.6

# JSON handling
orjson==3.9.10

# Async utilities
asyncio-mqtt==0.16.1

# Data validation
email-validator==2.1.0